chore(text-editor): move utils out of internal directory
