+++ src/components/text-editor/__internal__/index.ts
@@ -1,2 +1,1 @@
 export { default } from "./read-only-rte.component";
-export { createEmpty, createFromHTML } from "./utils";
